Improve PDF reader annotations sidebar interaction (#1321)
* Improve code

* Improve code

* Cleanup unused code

* Fix regression

* Propagate sidebar selected annotation change when filtering or searching

* Improve code

* Fix annotation sidebar editing selection update

* Improve code

* Fix annotation sidebar update when filter or search is used

* Fix PDF reader annotation update when color changes

* On PDF reader annotations update refresh filters for affected pages

* Improve code

* Improve code clarity

+++ Zotero/Scenes/Detail/PDF/Models/PDFReaderState.swift
@@ -16,9 +16,9 @@ typealias AnnotationDocumentLocation = (page: Int, boundingBox: CGRect)
 
 struct PDFReaderState: ViewModelState {
     struct Changes: OptionSet {
-        typealias RawValue = UInt32
+        typealias RawValue = UInt16
 
-        let rawValue: UInt32
+        let rawValue: UInt16
 
         static let annotations = Changes(rawValue: 1 << 0)
         static let selection = Changes(rawValue: 1 << 1)
